split cached_region_info_manager from observer
