+++ engine_tiflash/Cargo.toml
@@ -42,6 +42,7 @@ libc = "0.2"
 log_wrappers = { workspace = true }
 num_cpus = "1"
 online_config = { workspace = true }
+portable-atomic = "0.3"
 prometheus = { version = "0.13", features = ["nightly"] }
 prometheus-static-metric = "0.5"
 protobuf = "2"
